+++ src/basic/features/options/index.js
@@ -1,0 +1,2 @@
+// 🏪 옵션 서비스 exports
+export { optionService } from './optionService.js';
